Refactor yaml encoding to avoid __reduce__ errors
Documented in https://github.com/hgrecco/serialize/issues/7

serialize was overloading the object encoding/decoding, but didn't
properly differentiate itself from the built-in object handlers in
pyyaml. pyaml has a tagging facility to deal with this sort of thing, so
now serialize-handled classes are handled explicitly with their own
unique tag.

This is a breaking change, previously-serialized data will no longer
decode correctly.

This adds a register_class callback for formats, and yaml implements
that to register those classes with the built-in
serialize.yaml.Dumper/Loader for tagging/encoding.